requirements.txt
altair==5.2.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
eventregistry==9.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.14.0
fonttools==4.49.0
frozendict==2.4.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
html5lib==1.1
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.0.1
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multitasking==0.0.11
nest_asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parso==0.8.3
peewee==3.17.1
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.22
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.19.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
streamlit==1.31.1
sympy==1.12
tagui==1.50.0
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.2
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.41.2
wsproto==1.2.0
yfinance==0.2.37
zipp==3.17.0
